Code rework & animation
- moved dice and player components into a single component
- dice component animate on roll dice button click
- added comment to GameBoard for clarity & extracted a div to Index
- changed some colouring and UI
